refactor: make private rpc package to public and change timestamp format
I make private rpc which in the `internal` directory package to public
which in the `pkg` directory.

Also, I change the LogInfo's timestamp type time.Time to int64. For this
reason, I deal some methods to port it.
