+++ pkg/rpc/type.go
@@ -1,16 +1,12 @@
 package rpc
 
-import (
-	"time"
-)
-
 // LogInfo is part of LogMessage.
 // Consist of timestamp, filename, length.
 // timestamp is the time log created.
 // filename is the name of log file.
 // length is the size(bytes) per log message.
 type LogInfo struct {
-	Timestamp time.Time
+	Timestamp int64
 	Filename  string
 	Length    uint64
 }
